filterparent-773188: Exclude parent commands on input
When we have input then filter out the parent commands because we no
longer need the summary level that we need when we've not typed
anything.

This change breaks a number of tests because
SelectionType._findPredictions() uses it to match single commands. We
tried replacing the initial test for:

  // If the arg has a suffix then we're kind of 'done'. If we have an
  // exact match then we stop looking.
  if (arg.suffix.length > 0) {
    for (i = 0; i < lookup.length && predictions.length <
maxPredictions; i++) {
      option = lookup[i];
      if (option.name === arg.text) {
        this._addToPredictions(predictions, option, arg);
      }
    }

    if (predictions.length === 1) {
      return predictions;
    }
  }


But this broke other things.